convert class ShamirMnemonic to module shamir_mnemonic
fixes #7
